requirements.txt
PyMuPDF==1.23.26
pdf2image==1.17.0
pytesseract==0.3.10
Pillow==10.1.0